[BASIC] fix TI/TI$
Fixes #84

BASIC allocates dummy variables for TI and TI$ that point to
the "zero" constang in ROM. In the expression evaluator, it
is checked whether the variable points to any ROM location.
When FPLIB was separated out, the "is in ROM" check failed,
because it only allowed pointers in BASIC's ROM region, but
the zero constant is now contained in FPLIB.
The test was changed to allow any ROM location (i.e. >= $C000).