[KERNAL] Fix bank switching macros (#201)
* Fix banks.inc for joysticks.s

KVARS_START_TRASH_NZ and KVARS_END_TRASH_NZ are preserving A across the entire function call, in the one place where they're used this seems to be both unnecessary and incorrect, the A only needs to be preserved across the macro, and explicitly *should not* be preserved across the whole call. This seems to fix joystick polling, verified with the following test:

10?JOY(1):GOTO10
RUN

* Use slightly less expensive version of macro

Minor optimization to joystick_get, using a different KVARS_START macro that doesn't need to stow a value.