[BASIC] restore C flag in finh to fix null handling in GET/READ/INPUT
'chrget' will set the C flag to indicate it didn't get a digit (including when it got nothing at all).  'fin' relies on that when parsing a number.  The comparisons in 'finh' will set new values in C.  Since that code path is only followed when C=1, a 'sec' instruction fixes it.