[DOS] fixed writing LFN directory entries across sector boundary
When searching for an empty range of directory entries that would fit the LFN,
we may have to cross sector boundaries. The subroutine mark_dir_entry saves the
position of the start of a potential empty range, and if the range is found to
be large enough, the subroutine rewind_dir_entry restores the state, so we can
start writing directory entries there.

These functions save the current cluster number, the current LBA and the
directory entry position within the sector, but they fail to save the number of
the sector within the cluster.

In the example of two sectors per cluster and the empty range spanning two
sectors within a cluster, a call to next_sector (while searching for the empty
range) increments the sector within the cluster (from 0 to 1). Then we rewind.
When finally writing the LFN entries, we call next_sector again after a while,
but since "sector within the cluster" is already 1, it'll be incremented to 2,
triggering a load of the following cluster, which does not exist, so we can't
write the remaining entries.

This fix saves "sector within the cluster" as well.

Fixes #224