[KERNAL] Move graph and font (#334)
This PR moves the graph and fonts functions from ROM bank 0 to a separate ROM bank (8). The test demo code is moved from ROM bank 8 to ROM bank 9. The default framebuffer driver remains in ROM bank 0.

The move is made to free up space in ROM bank 0, which is needed for future development of the Kernal, as described by @mist64 in issue https://github.com/commanderx16/x16-rom/issues/316. This PR frees up about 5.5 kB in ROM bank 0.

The ROM bank 0 vectors for graph and font functions are still there not to break any existing code. These bank 0 functions are, however, only wrappers that call code in bank 8.

Multiple bank switching has a negative effect on performance, and there is some performance to gain from calling the bank 8 functions directly instead of through the bank 0 vectors. I'm not too worried about performance though, as I don't think those who seek performance will be using the built-in graph and fonts library.

At the beginning of ROM bank 8, there's a jump table for all externally exposed functions. This jump table is available in the include file inc/graphics.inc.

The graph and font functions need to call a few Kernal functions and use a few Kernal RAM variables that have no fixed address. These symbols and their memory addresses are determined during the link stage by the utility "findsymbols" that I created. This utility retrieves memory addresses of symbols from a .sym file. The utility is a simple bash script that invokes awk. Thanks to @ZeroByteOrg for suggesting this solution.

In the current master branch there is a bug that affects the primary foreground color setting for graph functions. The primary color is held in RAM variable col1, but this memory address is also used by the Kernal for a different purpose (activity_led_state). This bug has been fixed in the PR.