memory_decompress can decompress to VRAM
decompression to VRAM did not work, because during decompression reads from already decompressed bytes (backreferences) are also needed.

To make it work I had to do the following changes:

1) the code implicitly assumes that you're decompressing to VERA ram and that the dataport 0 is the destination address.

2) had to make  add_match offset different for RAM and IO. The IO version of it now reads the current address from vera ADDR_L, M and H (on data port 0) does a 17 bit add and stores the calculated address again to vera, but on data port 1

3) had to make copy_backreference different for RAM and IO. The io variant reads from data1 and writes to data0.

With all these changes I then had 4 "function pointers" and needed to push 3 additional 16 bit registers onto the stack, and I didn't want that, so I changed the code to not use function pointers but to do the dispatch in each function based on the io bank - this is 2 cycles slower than the original code, but it results in smaller code, and since I now also use it for the putdst method  I actually now need to push 1 word less onto the stack.