Enable kernal to accurately detect amount of memory installed (#8)
* Enable kernal to accurately detect amount of memory installed
ROM was relying on the fact that the emulator wraps around to
bank 0 when no more banks of memory are present. Real hardware
does not have this functionality instead, writes to empty banks
are not stored anywhere. For this reason I have added a bit of
code that checks that a write to a bank can actually be read
back, then it compares it with value on bank 0 to ensure
continued support in the emulator as well.
I have tested the ROM in latest emulator and on unoffical
hardware designed by Joe Burks (Wavicle)

* Added comment explaining ASL to INC switch.
Switched to INC instead of ASL to ensure that odd memory
sizes such as 1536K are also detected. With ASL it would
only be possible to detect 1, 2, 4, 8, 16, 32, 64, 128 or
256 banks which means 1536K could not be detected.

* Complete rewrite of ram bank detection code.
The code has been rewritten to do read, xor, store, compare
on two address in a bank as well as compare with bank 0 to
check for wrap-around.

* Another rewrite as suggested by stefan-b-jakobsson
Original bank-detector is kept, almost intact, but another
loop is added to ensure that the correct amount of memory
is detected when there is no wrap-around to bank0

* RAM bank does not need to bet set to 0
Realized that a few lines later, RAM bank is set to 1

---------

Co-authored-by: Jimmy Dansbo <jimmy@dansbo.dk>